+++ config/test.exs
@@ -1,0 +1,6 @@
+use Mix.Config
+
+config :junit_formatter,
+  print_report_file: true
+
+config :logger, :level, :info