+++ test/test_helper.exs
@@ -1,2 +1,2 @@
 ExUnit.configure formatters: [JUnitFormatter, ExUnit.CLIFormatter]
-ExUnit.start [trace: "--trace" in System.argv]
+ExUnit.start [trace: "--trace" in System.argv, timeout: 5_000]